requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
cached-property==1.5.1
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
Django==2.1.8
django-bootstrap-pagination==1.7.0
django-crispy-forms==1.7.2
djangorestframework==3.9.4
docopt==0.6.2
docutils==0.14
future==0.16.0
idna==2.7
jmespath==0.9.4
joblib==0.13.2
jsonschema==2.6.0
pathspec==0.5.9
pyasn1==0.4.5
python-dateutil==2.8.0
pytz==2019.1
PyYAML==3.13
requests==2.20.1
rsa==3.4.2
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.1
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.7
websocket-client==0.56.0
